Rename the lambda function in order to become a form-submission triggered one
